refactor: changes as per. PR comments
