+++ src/modules/Auth/Auth.middleware.test.ts
@@ -33,7 +33,7 @@ import {
 
 describe('modules/Auth/Auth.middleware', () => {
   describe('authLoginFlow', () => {
-    it('should correctly handle being called', async () => {
+    it('should correctly handle being called', () => {
       // given ... the authLogin action is fired
       const create = createMiddlewareMock(jest)
       const credentials = {
@@ -45,13 +45,13 @@ describe('modules/Auth/Auth.middleware', () => {
       const { invoke, next, store } = create(SUT.authLoginFlow)
 
       // when ... we respond to the authLogin action
-      await invoke(action)
+      invoke(action)
 
       // then ... the login API should be called
       expect(next).toHaveBeenCalledWith(action)
       expect(store.dispatch).toHaveBeenCalled()
     })
-    it('should correctly login the user in', async () => {
+    it('should correctly login the user in', () => {
       // given ... the authLogin action is fired
       const create = createMiddlewareMock(jest)
       const credentials = {
@@ -63,7 +63,7 @@ describe('modules/Auth/Auth.middleware', () => {
       const { invoke, store } = create(SUT.authLoginFlow)
 
       // when ... we respond to the authLogin action
-      await invoke(action)
+      invoke(action)
 
       // then ... the login API should be called
       expect(store.dispatch).toHaveBeenCalledWith(
@@ -79,29 +79,31 @@ describe('modules/Auth/Auth.middleware', () => {
     })
   })
   describe('authSocialLoginFlow', () => {
-    it('should correctly handle being called', async () => {
+    it('should correctly handle being called', () => {
       // given ... the action is fired
       const create = createMiddlewareMock(jest)
       const mockSocialLogin = jest.fn().mockResolvedValue(true)
       const action = authSocialRegistration('provider')
       // @ts-ignore
       const { invoke, next } = create(SUT.authSocialRegistrationFlow({ socialAuth: mockSocialLogin }))
-      await invoke(action)
+      invoke(action)
 
       expect(next).toHaveBeenCalledWith(action)
     })
-    it('should provide feedback on login failure', async () => {
+    it('should provide feedback on login failure', () => {
       // given ... the action is fired
       const create = createMiddlewareMock(jest)
-      const mockSocialLogin = jest.fn().mockRejectedValue(true)
+      const mockSocialLogin = jest.fn().mockImplementation(() => {
+        throw new Error('FAILURE')
+      })
       const action = authSocialLogin('provider')
       // @ts-ignore
       const { store, invoke } = create(SUT.authSocialLoginFlow({ socialAuth: mockSocialLogin }))
-      await invoke(action)
+      invoke(action)
 
       expect(store.dispatch).toHaveBeenCalledWith(authSocialLoginFailure)
     })
-    it('should correctly get user data from the provider', async () => {
+    it('should correctly get user data from the provider', () => {
       const create = createMiddlewareMock(jest)
       const mockNotification = jest.fn()
       const mockSocialLogin = jest.fn().mockResolvedValue(true)
@@ -111,29 +113,29 @@ describe('modules/Auth/Auth.middleware', () => {
       const { invoke } = create(
         SUT.authSocialLoginFlow({ socialAuth: mockSocialLogin, notification: mockNotification }),
       )
-      await invoke(action)
+      invoke(action)
 
       expect(mockSocialLogin).toHaveBeenCalled()
     })
   })
   describe('authSocialLoginSuccessFlow', () => {
-    it('should correctly handle being called', async () => {
+    it('should correctly handle being called', () => {
       const create = createMiddlewareMock(jest)
       // given ... the action is fired
       const action = authSocialLoginSuccess(userSocialRegistrationData)
       // @ts-ignore
       const { invoke, next } = create(SUT.authSocialLoginSuccessFlow)
-      await invoke(action)
+      invoke(action)
 
       expect(next).toHaveBeenCalledWith(action)
     })
-    it('should correctly login the user', async () => {
+    it('should correctly login the user', () => {
       // given ... the action is fired
       const create = createMiddlewareMock(jest)
       const action = authSocialLoginSuccess(userSocialLoginData)
       // @ts-ignore
       const { invoke, store } = create(SUT.authSocialLoginSuccessFlow)
-      await invoke(action)
+      invoke(action)
 
       expect(store.dispatch).toHaveBeenCalledWith(
         ApiActions.apiRequest(
@@ -148,58 +150,60 @@ describe('modules/Auth/Auth.middleware', () => {
     })
   })
   describe('authSocialRegistrationFlow', () => {
-    it('should correctly handle being called', async () => {
+    it('should correctly handle being called', () => {
       // given ... the action is fired
       const create = createMiddlewareMock(jest)
       const mockSocialRegistration = jest.fn().mockResolvedValue(true)
       const action = authSocialRegistration('provider')
       // @ts-ignore
       const { invoke, next } = create(SUT.authSocialRegistrationFlow({ socialAuth: mockSocialRegistration }))
-      await invoke(action)
+      invoke(action)
 
       expect(next).toHaveBeenCalledWith(action)
     })
-    it('should provide feedback on registration failure', async () => {
+    it('should provide feedback on registration failure', () => {
       // given ... the action is fired
       const create = createMiddlewareMock(jest)
-      const mockSocialRegistration = jest.fn().mockRejectedValue(true)
+      const mockSocialRegistration = jest.fn().mockImplementation(() => {
+        throw new Error('FAILURE')
+      })
       const action = authSocialRegistration('provider')
       // @ts-ignore
       const { invoke, store } = create(SUT.authSocialRegistrationFlow({ socialAuth: mockSocialRegistration }))
-      await invoke(action)
+      invoke(action)
 
       expect(store.dispatch).toHaveBeenCalledWith(authSocialRegistrationFailure)
     })
-    it('should correctly get user data from the provider', async () => {
+    it('should correctly get user data from the provider', () => {
       const create = createMiddlewareMock(jest)
       const mockSocialRegistration = jest.fn().mockResolvedValue(true)
       // given ... the action is fired
       const action = authSocialRegistration('provider')
       // @ts-ignore
       const { invoke } = create(SUT.authSocialRegistrationFlow({ socialAuth: mockSocialRegistration }))
-      await invoke(action)
+      invoke(action)
 
       expect(mockSocialRegistration).toHaveBeenCalled()
     })
   })
   describe('authSocialRegistrationSuccessFlow', () => {
-    it('should correctly handle being called', async () => {
+    it('should correctly handle being called', () => {
       const create = createMiddlewareMock(jest)
       // given ... the action is fired
       const action = authSocialRegistrationSuccess(userSocialRegistrationData)
       // @ts-ignore
       const { invoke, next } = create(SUT.authSocialRegistrationSuccessFlow)
-      await invoke(action)
+      invoke(action)
 
       expect(next).toHaveBeenCalledWith(action)
     })
-    it('should correctly register the user', async () => {
+    it('should correctly register the user', () => {
       // given ... the action is fired
       const create = createMiddlewareMock(jest)
       const action = authSocialRegistrationSuccess(userSocialRegistrationData)
       // @ts-ignore
       const { invoke, store } = create(SUT.authSocialRegistrationSuccessFlow)
-      await invoke(action)
+      invoke(action)
 
       expect(store.dispatch).toHaveBeenCalledWith(
         ApiActions.apiRequest(
@@ -214,7 +218,7 @@ describe('modules/Auth/Auth.middleware', () => {
     })
   })
   describe('authLoginSuccessFlow', () => {
-    it('should correctly catch a successful login action', async () => {
+    it('should correctly catch a successful login action', () => {
       // given ... the authLoginSuccess action is fired
       const create = createMiddlewareMock(jest)
       const action = authLoginSuccess(defaultUserLoginResponseData)
@@ -223,13 +227,13 @@ describe('modules/Auth/Auth.middleware', () => {
       const { store, invoke, next } = create(SUT.authLoginSuccessFlow({ notification: mockNotification }))
 
       // when ... we respond to the authLoginSuccess action
-      await invoke(action)
+      invoke(action)
 
       // then ... we should correctly catch the action
       expect(next).toHaveBeenCalledWith(action)
       expect(store.dispatch).toHaveBeenCalled()
     })
-    it('should correctly send a notification to the user', async () => {
+    it('should correctly send a notification to the user', () => {
       // given ... the authLogin action is fired
       const create = createMiddlewareMock(jest)
       // @ts-ignore
@@ -239,12 +243,12 @@ describe('modules/Auth/Auth.middleware', () => {
       const { invoke } = create(SUT.authLoginSuccessFlow({ notification: mockNotification }))
 
       // when ... we respond to the authLogin action
-      await invoke(action)
+      invoke(action)
 
       // then ... the login API should be called
       expect(mockNotification).toHaveBeenCalled()
     })
-    it('should extract and transmit auth credentials and refresh token', async () => {
+    it('should extract and transmit auth credentials and refresh token', () => {
       // given ... the authLogin action is fired
       const create = createMiddlewareMock(jest)
       // @ts-ignore
@@ -254,7 +258,7 @@ describe('modules/Auth/Auth.middleware', () => {
       const { store, invoke } = create(SUT.authLoginSuccessFlow({ notification: mockNotification }))
 
       // when ... we respond to the authLogin action
-      await invoke(action)
+      invoke(action)
 
       // then ... the login API should be called
       expect(store.dispatch).toHaveBeenCalledWith(setAuthCredentials({ token: 'USER_TOKEN', expiresAt: 'EXPIRY_DATE' }))
@@ -262,7 +266,7 @@ describe('modules/Auth/Auth.middleware', () => {
     })
   })
   describe('setSecureRefreshTokenFlow', () => {
-    it('should correctly handle being called', async () => {
+    it('should correctly handle being called', () => {
       // given ... the authLoginSuccess action is fired
       const create = createMiddlewareMock(jest)
       const action = setSecureRefreshToken('REFRESH_TOKEN')
@@ -271,7 +275,7 @@ describe('modules/Auth/Auth.middleware', () => {
       const { next, invoke } = create(SUT.setSecureRefreshTokenFlow(setSecureItemStub))
 
       // when ... we respond to the authLoginSuccess action
-      await invoke(action)
+      invoke(action)
 
       // then ...
       // ... we should correctly attempt to store the token
@@ -313,7 +317,7 @@ describe('modules/Auth/Auth.middleware', () => {
     })
   })
   describe('authLoginFailureFlow', () => {
-    it('should correctly send a notification to the user', async () => {
+    it('should correctly send a notification to the user', () => {
       // given ... the authLogin action is fired
       const create = createMiddlewareMock(jest)
       // @ts-ignore
@@ -324,14 +328,14 @@ describe('modules/Auth/Auth.middleware', () => {
       const { invoke } = create(SUT.authLoginFailureFlow({ notification: mockNotification }))
 
       // when ... we respond to the authLogin action
-      await invoke(action)
+      invoke(action)
 
       // then ... the login API should be called
       expect(mockNotification).toHaveBeenCalled()
     })
   })
   describe('authLogoutFlow', () => {
-    it('should correctly logout the user', async () => {
+    it('should correctly logout the user', () => {
       // given ... the authLogout action is fired
       const create = createMiddlewareMock(jest)
       const action = authLogout()
@@ -339,14 +343,14 @@ describe('modules/Auth/Auth.middleware', () => {
       const { store, invoke } = create(SUT.authLogoutFlow)
 
       // when ... we respond to the authLogout action
-      await invoke(action)
+      invoke(action)
 
       // then ... the reset APP should be called
       expect(store.dispatch).toHaveBeenCalledWith(resetAppData())
     })
   })
   describe('authRegistrationFlow', () => {
-    it('should ignore other actions', async () => {
+    it('should ignore other actions', () => {
       // given ... an action is fired
       const create = createMiddlewareMock(jest)
       const mockApi = jest.fn()
@@ -355,14 +359,14 @@ describe('modules/Auth/Auth.middleware', () => {
       const { invoke, next } = create(SUT.authRegistrationFlow)
 
       // when ... we invoke another action
-      await invoke(action)
+      invoke(action)
 
       // then ... next should have been called with the action
       expect(next).toHaveBeenCalledWith(action)
       expect(mockApi).not.toHaveBeenCalled()
     })
 
-    it('should correctly handle registration', async () => {
+    it('should correctly handle registration', () => {
       // given ... the authRegistration action is fired
       const create = createMiddlewareMock(jest)
       const action = authRegistration(userRegistrationData)
@@ -370,13 +374,13 @@ describe('modules/Auth/Auth.middleware', () => {
       const { invoke, next } = create(SUT.authRegistrationFlow)
 
       // when ... we respond to the authRegistration action
-      await invoke(action)
+      invoke(action)
 
       // then ... the register API should be called
       expect(next).toHaveBeenCalledWith(action)
     })
 
-    it('should correctly register a user', async () => {
+    it('should correctly register a user', () => {
       // given ... the register api is called and returns successfully
       const create = createMiddlewareMock(jest)
       const action = authRegistration(userRegistrationData)
@@ -385,7 +389,7 @@ describe('modules/Auth/Auth.middleware', () => {
       const { store, invoke } = create(SUT.authRegistrationFlow)
 
       // when ... we respond to the successful state
-      await invoke(action)
+      invoke(action)
 
       // then ... the success action should be dispatched with the response
       expect(store.dispatch).toHaveBeenCalledWith(
@@ -401,7 +405,7 @@ describe('modules/Auth/Auth.middleware', () => {
     })
   })
   describe('authRegistrationSuccessFlow', () => {
-    it('should correctly send a notification to the user', async () => {
+    it('should correctly send a notification to the user', () => {
       // middleware is dependent on a populated state
       const state = {
         auth: {
@@ -418,14 +422,14 @@ describe('modules/Auth/Auth.middleware', () => {
       // @ts-ignore
       const { invoke } = create(SUT.authRegistrationSuccessFlow({ notification: mockNotification }))
 
-      await invoke(action)
+      invoke(action)
 
       // then ... the notification should be calleds
       expect(mockNotification).toHaveBeenCalled()
     })
   })
   describe('authRegistrationFailureFlow', () => {
-    it('should correctly send a notification to the user', async () => {
+    it('should correctly send a notification to the user', () => {
       // given ... the authRegistration action is fired
       const create = createMiddlewareMock(jest)
       // @ts-ignore
@@ -436,14 +440,14 @@ describe('modules/Auth/Auth.middleware', () => {
       const { invoke } = create(SUT.authRegistrationFailureFlow({ notification: mockNotification }))
 
       // when ... we respond to the authRegistration action
-      await invoke(action)
+      invoke(action)
 
       // then ... the notification should be called
       expect(mockNotification).toHaveBeenCalled()
     })
   })
   describe('authSocialLoginFailureFlow', () => {
-    it('should correctly send a notification to the user', async () => {
+    it('should correctly send a notification to the user', () => {
       // given ... the authRegistration action is fired
       const create = createMiddlewareMock(jest)
       // @ts-ignore
@@ -454,14 +458,14 @@ describe('modules/Auth/Auth.middleware', () => {
       const { invoke } = create(SUT.authSocialLoginFailureFlow({ notification: mockNotification }))
 
       // when ... we respond to the authRegistration action
-      await invoke(action)
+      invoke(action)
 
       // then ... the notification should be called
       expect(mockNotification).toHaveBeenCalled()
     })
   })
   describe('authSocialRegistrationFailureFlow', () => {
-    it('should correctly send a notification to the user', async () => {
+    it('should correctly send a notification to the user', () => {
       // given ... the authRegistration action is fired
       const create = createMiddlewareMock(jest)
       // @ts-ignore
@@ -472,7 +476,7 @@ describe('modules/Auth/Auth.middleware', () => {
       const { invoke } = create(SUT.authSocialRegistrationFailureFlow({ notification: mockNotification }))
 
       // when ... we respond to the authRegistration action
-      await invoke(action)
+      invoke(action)
 
       // then ... the notification should be called
       expect(mockNotification).toHaveBeenCalled()
